イベントディスパッチャを分離
