+++ Packages/com.nomlas.topazchat/Runtime/Scripts/Player.cs
@@ -6,7 +6,7 @@ using VRC.SDKBase;
 
 namespace Nomlas.TopazChat
 {
-    public class Player : TopazChatBase
+    public class Player : EventDispatcher
     {
         #region Inspector
         [SerializeField] internal VRCUrl defaultStreamURL;
@@ -55,59 +55,6 @@ namespace Nomlas.TopazChat
 
         public Material ScreenMaterial { get => screen.sharedMaterial; }
 
-        #region Listener
-        /// <summary>
-        /// イベントリスナーの配列。
-        /// AddEventListenerされるまではnullになっていることに注意してください。
-        /// </summary>
-        private PlayerEventListener[] listeners;
-
-        internal void AddEventListener(PlayerEventListener listener)
-        {
-            if (listener == null)
-            {
-                LogError("空のEventListenerが渡されました");
-                return;
-            }
-            if (listeners == null)
-                listeners = new PlayerEventListener[0];
-            var array = new PlayerEventListener[listeners.Length + 1];
-            listeners.CopyTo(array, 0);
-            array[listeners.Length] = listener;
-            listeners = array;
-            Log("Added EventListener");
-            listener.OnListenerReady();
-        }
-
-        internal void ShowMessage(string msg)
-        {
-            _ShowMessage(msg, MessageLevel.Info);
-        }
-
-        internal void ShowMessage(string msg, MessageLevel level)
-        {
-            _ShowMessage(msg, level);
-        }
-
-        private void _ShowMessage(string msg, MessageLevel level)
-        {
-            if (!Utilities.IsValid(listeners)) return;
-            for (int i = 0; i < listeners.Length; i++)
-            {
-                listeners[i].UpdateMessage($"<color={TopazUtils.MessageLevelColor(level)}>{msg}</color>");
-            }
-        }
-
-        private void UpdatePlayerStatus(PlayerStatus playerStatus)
-        {
-            if (!Utilities.IsValid(listeners)) return;
-            for (int i = 0; i < listeners.Length; i++)
-            {
-                listeners[i].UpdateStatus(playerStatus);
-            }
-        }
-        #endregion
-
         /// <summary>
         /// 指定したURLで再生します
         /// </summary>
@@ -147,15 +94,6 @@ namespace Nomlas.TopazChat
             else
             {
                 PlayURL(url, PlayType.Play);
-            }
-        }
-
-        private void UpdateURL(VRCUrl url, VRCUrl url_Android)
-        {
-            if (!Utilities.IsValid(listeners)) return;
-            for (int i = 0; i < listeners.Length; i++)
-            {
-                listeners[i].UpdateURL(url, url_Android);
             }
         }
